web/requirements.txt
flask>=3.0
pyyaml>=6.0
pillow>=10.0
